add ssh password prompt and summary report
