+++ scripts/data_migration.py
@@ -4,9 +4,10 @@ import os
 import sys
 import argparse
 import re
-from solr_connection import SolrInputSource, SolrOutputSink, SSHConnection, LocalConnection
+from solr_endpoint import SolrInputSource, SolrOutputSink, LocalConnection
+from ssh_connection import SSHConnection
 from transporter import Transporter, DataAnonymizer
-from file_connection import FileInputSource, FileOutputSink
+from file_endpoint import FileInputSource, FileOutputSink
 import requests
 import tempfile
 import zipfile
@@ -40,8 +41,6 @@ def main():
     parser.add_argument("destination", help="Data destination - can be a filepath (/tmp/blah), " +
                                             "the local Cb server (local), or a remote Cb server (root@cb5.server:2202)")
     parser.add_argument("-v", "--verbose", help="Increase output verbosity", action="store_true")
-    parser.add_argument("--key", help="SSH private key location", action="store",
-                        default=os.path.join(os.path.expanduser('~'), '.ssh', 'id_rsa'))
     parser.add_argument("--anonymize", help="Anonymize data in transport", action="store_true", default=False)
     parser.add_argument("-q", "--query", help="Source data query (required for server input)", action="store")
     parser.add_argument("--tree", help="Traverse up and down process tree", action="store_true", default=False)
@@ -90,7 +89,7 @@ def main():
             return 2
 
         input_connection = SSHConnection(username=source_parts.group(1), hostname=source_parts.group(2),
-                                         port=port_number, private_key=options.key)
+                                         port=port_number)
         input_source = SolrInputSource(input_connection, query=options.query)
     else:
         # source_parts is a file path
@@ -104,7 +103,7 @@ def main():
         if destination_parts.group(4):
             port_number = int(destination_parts.group(4))
         output_connection = SSHConnection(username=destination_parts.group(1), hostname=destination_parts.group(2),
-                                          port=port_number, private_key=options.key)
+                                          port=port_number)
         output_sink = SolrOutputSink(output_connection)
     else:
         output_sink = FileOutputSink(options.destination)
@@ -115,6 +114,8 @@ def main():
         t.add_anonymizer(DataAnonymizer())
 
     t.transport(debug=options.verbose)
+    print "Migration complete!"
+    print t.get_report()
 
 # TODO: add exception wrapper so we clean up after ourselves if there's an error
 if __name__ == '__main__':
